feat: Add table reader
