+++ docx-core/src/documents/elements/table.rs
@@ -105,4 +105,13 @@ mod tests {
 </w:tblGrid><w:tr><w:trPr /></w:tr></w:tbl>"#
         );
     }
+
+    #[test]
+    fn test_table_json() {
+        let t = Table::new(vec![]).set_grid(vec![100, 200, 300]);
+        assert_eq!(
+            serde_json::to_string(&t).unwrap(),
+            r#"{"rows":[],"grid":[100,200,300],"hasNumbering":false,"property":{"width":{"width":0,"widthType":"Auto"},"justification":"left","borders":{"top":{"position":"top","borderType":"single","size":2,"space":0,"color":"000000"},"left":{"position":"left","borderType":"single","size":2,"space":0,"color":"000000"},"bottom":{"position":"bottom","borderType":"single","size":2,"space":0,"color":"000000"},"right":{"position":"right","borderType":"single","size":2,"space":0,"color":"000000"},"insideH":{"position":"insideH","borderType":"single","size":2,"space":0,"color":"000000"},"insideV":{"position":"insideV","borderType":"single","size":2,"space":0,"color":"000000"}},"margins":{"top":55,"left":54,"bottom":55,"right":55},"indent":null}}"#
+        );
+    }
 }
