+++ docx-core/src/documents/elements/table_borders.rs
@@ -60,10 +60,10 @@ impl BuildXML for TableBorder {
             BorderPosition::Right => {
                 base.border_right(self.border_type, self.size, self.space, &self.color)
             }
-            BorderPosition::IndideH => {
+            BorderPosition::InsideH => {
                 base.border_inside_h(self.border_type, self.size, self.space, &self.color)
             }
-            BorderPosition::IndideV => {
+            BorderPosition::InsideV => {
                 base.border_inside_v(self.border_type, self.size, self.space, &self.color)
             }
         };
@@ -101,8 +101,8 @@ impl Default for TableBorders {
             left: Some(TableBorder::new(BorderPosition::Left)),
             bottom: Some(TableBorder::new(BorderPosition::Bottom)),
             right: Some(TableBorder::new(BorderPosition::Right)),
-            inside_h: Some(TableBorder::new(BorderPosition::IndideH)),
-            inside_v: Some(TableBorder::new(BorderPosition::IndideV)),
+            inside_h: Some(TableBorder::new(BorderPosition::InsideH)),
+            inside_v: Some(TableBorder::new(BorderPosition::InsideV)),
         }
     }
 }
@@ -118,8 +118,8 @@ impl TableBorders {
             BorderPosition::Left => self.left = Some(border),
             BorderPosition::Bottom => self.bottom = Some(border),
             BorderPosition::Right => self.right = Some(border),
-            BorderPosition::IndideH => self.inside_h = Some(border),
-            BorderPosition::IndideV => self.inside_v = Some(border),
+            BorderPosition::InsideH => self.inside_h = Some(border),
+            BorderPosition::InsideV => self.inside_v = Some(border),
         };
         self
     }
@@ -130,8 +130,8 @@ impl TableBorders {
             BorderPosition::Left => self.left = None,
             BorderPosition::Bottom => self.bottom = None,
             BorderPosition::Right => self.right = None,
-            BorderPosition::IndideH => self.inside_h = None,
-            BorderPosition::IndideV => self.inside_v = None,
+            BorderPosition::InsideH => self.inside_h = None,
+            BorderPosition::InsideV => self.inside_v = None,
         };
         self
     }
