+++ docx-core/src/documents/rels.rs
@@ -1,14 +1,11 @@
 use serde::{Deserialize, Serialize};
-use std::io::Read;
-use xml::reader::{EventReader, XmlEvent};
 
 use crate::documents::BuildXML;
-use crate::reader::{FromXML, ReaderError};
 use crate::xml_builder::*;
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 pub struct Rels {
-    rels: Vec<(String, String, String)>,
+    pub rels: Vec<(String, String, String)>,
 }
 
 impl Rels {
@@ -70,43 +67,6 @@ impl BuildXML for Rels {
     }
 }
 
-impl FromXML for Rels {
-    fn from_xml<R: Read>(reader: R) -> Result<Self, ReaderError> {
-        let parser = EventReader::new(reader);
-        let mut s = Self::default();
-        let mut depth = 0;
-        for e in parser {
-            match e {
-                Ok(XmlEvent::StartElement { attributes, .. }) => {
-                    if depth == 1 {
-                        let mut id = "".to_owned();
-                        let mut rel_type = "".to_owned();
-                        let mut target = "".to_owned();
-                        for attr in attributes {
-                            let name: &str = &attr.name.local_name;
-                            if name == "Id" {
-                                id = attr.value.clone();
-                            } else if name == "Type" {
-                                rel_type = attr.value.clone();
-                            } else if name == "Target" {
-                                target = attr.value.clone();
-                            }
-                        }
-                        s = s.add_rel(id, rel_type, target);
-                    }
-                    depth += 1;
-                }
-                Ok(XmlEvent::EndElement { .. }) => {
-                    depth -= 1;
-                }
-                Err(_) => return Err(ReaderError::XMLReadError),
-                _ => {}
-            }
-        }
-        Ok(s)
-    }
-}
-
 #[cfg(test)]
 mod tests {
 
@@ -128,22 +88,5 @@ mod tests {
   <Relationship Id="rId3" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="word/document.xml" />
 </Relationships>"#
         );
-    }
-
-    #[test]
-    fn test_from_xml() {
-        let xml = r#"<?xml version="1.0" encoding="UTF-8"?>
-<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">
-  <Relationship Id="rId1" Type="http://schemas.openxmlformats.org/package/2006/relationships/metadata/core-properties" Target="docProps/core.xml" />
-</Relationships>"#;
-        let c = Rels::from_xml(xml.as_bytes()).unwrap();
-        let mut rels = Vec::new();
-        rels.push((
-            "http://schemas.openxmlformats.org/package/2006/relationships/metadata/core-properties"
-                .to_owned(),
-            "rId1".to_owned(),
-            "docProps/core.xml".to_owned(),
-        ));
-        assert_eq!(Rels { rels }, c);
     }
 }
